requirements.txt
flask
transformers
torch
torchvision
easyocr
requests
beautifulsoup4
numpy
pillow
gunicorn